user-auth/requirements.txt
dnspython==1.16.0
Flask==1.0.3
Flask-Login==0.5.0
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
gcs-oauth2-boto-plugin==2.3
google-api-core==1.16.0
google-apitools==0.5.28
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-reauth==0.1.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
GoogleAppEngineCloudStorageClient==1.9.22.1
gsutil==4.34
gunicorn==19.9.0
h5py==2.9.0
httplib2==0.12.3
mongoengine==0.19.1
numpy==1.16.3
passlib==1.7.2
PyJWT==1.7.1
pymongo==3.10.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pytest==4.6.3
python-dateutil==2.8.0
python-gflags==3.1.2
python-nginx==1.5.3
PyYAML==5.1
requests==2.21.0
scipy==1.2.1
six==1.12.0
urllib3==1.24.2
uWSGI==2.0.18
webencodings==0.5.1
websocket-client==0.48.0
Werkzeug==0.15.4
WTForms==2.2.1